requirements.txt
opencv-python
sockets
websocket-client
rel
ffmpeg-python
opencv-python